ticket: feature(dashboard) move files and created TopWidget
