+++ client/dashboard/src/components/App/App.tsx
@@ -1,17 +1,14 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import * as S from "./App.css";
-import { PATHS } from "../../../utilities/constants";
+import { PATHS } from "../../utilities/constants";
+import TopWidgets from "../TopWidgets";
 
 const App: React.FC = () => {
   return (
     <S.Main>
       <BrowserRouter>
         <>
-          <Routes>
-            <Route path={PATHS.HOME}>
-              <Route index element={<div>Home Page</div>} />
-            </Route>
-          </Routes>
+          <TopWidgets />
         </>
       </BrowserRouter>
     </S.Main>
