+++ client/dashboard/src/components/TopWidgets/TopWidgets.css.ts
@@ -1,0 +1,9 @@
+import { Grid, styled } from "@mui/material";
+import { StyledComponent } from "@emotion/styled";
+
+export const Container: StyledComponent<any, any, any> = styled(Grid)`
+  margin-left: 24px;
+  margin-top: 60px;
+  flex-direction: column;
+  background-color: inherit;
+`;
