+++ client/dashboard/src/components/TopWidgets/TopWidgets.tsx
@@ -1,0 +1,14 @@
+import { Grid } from "@mui/material";
+import * as S from "./TopWidgets.css";
+
+const TopWidgets = () => {
+  return (
+    <S.Container container maxWidth="md">
+      <Grid container item justifyContent="space-between">
+        TopWidgets
+      </Grid>
+    </S.Container>
+  );
+};
+
+export default TopWidgets;
