+++ client/dashboard/src/components/TopWidgets/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TopWidgets";
