+++ client/dashboard/src/zhabskyi-dashboard-app.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import singleSpaReact from "single-spa-react";
 
-import Root from "./components/base/Root";
+import Root from "./components/Root";
 
 const lifecycles = singleSpaReact({
   React,
