+++ client/orchestrator/src/index.ejs
@@ -49,6 +49,7 @@
       {
         "imports": {
           "@Zhabskyi/root-config": "//localhost:3000/zhabskyi-root-config.js",
+          "@Zhabskyi/dashboard": "//localhost:3001/zhabskyi-dashboard-app.js",
           "react": "https://cdn.jsdelivr.net/npm/react@18.3.1/umd/react.production.min.js",
           "single-spa": "https://cdn.jsdelivr.net/npm/single-spa@5.9.0/lib/system/single-spa.min.js",
           "react-dom": "https://cdn.jsdelivr.net/npm/react-dom@18.3.1/umd/react-dom.production.min.js"
